Closes #47 (#48)
* Moved before outside the spec file

* Refactoring

* Refactoring further

* Removed unused variables
